@@ -20,51 +20,51 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^5.2.6",
-    "@angular/cdk": "^5.2.2",
-    "@angular/common": "^5.2.6",
-    "@angular/compiler": "^5.2.6",
-    "@angular/core": "^5.2.6",
-    "@angular/flex-layout": "^5.0.0-beta.13",
-    "@angular/forms": "^5.2.6",
-    "@angular/http": "^5.2.6",
-    "@angular/material": "^5.2.2",
-    "@angular/platform-browser": "^5.2.6",
-    "@angular/platform-browser-dynamic": "^5.2.6",
-    "@angular/router": "^5.2.6",
-    "@ngrx/store": "^5.1.0",
+    "@angular/animations": "^5.2.10",
+    "@angular/cdk": "^5.2.5",
+    "@angular/common": "^5.2.10",
+    "@angular/compiler": "^5.2.10",
+    "@angular/core": "^5.2.10",
+    "@angular/flex-layout": "^5.0.0-beta.14",
+    "@angular/forms": "^5.2.10",
+    "@angular/http": "^5.2.10",
+    "@angular/material": "^5.2.5",
+    "@angular/platform-browser": "^5.2.10",
+    "@angular/platform-browser-dynamic": "^5.2.10",
+    "@angular/router": "^5.2.10",
+    "@ngrx/store": "^5.2.0",
     "angularfire2": "^5.0.0-rc.5-next",
-    "core-js": "^2.4.1",
-    "firebase": "^4.10.1",
+    "core-js": "^2.5.5",
+    "firebase": "^4.13.1",
     "hammerjs": "^2.0.8",
     "rxjs": "^5.5.6",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.6",
-    "@angular/language-service": "^5.2.6",
+    "@angular/cli": "^1.7.4",
+    "@angular/compiler-cli": "^5.2.10",
+    "@angular/language-service": "^5.2.10",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.101",
-    "codelyzer": "^4.0.1",
+    "@types/node": "^9.6.6",
+    "codelyzer": "^4.3.0",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~2.0.2",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "prettier": "1.11.1",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "prettier": "^1.12.1",
     "pretty-quick": "^1.4.1",
-    "protractor": "~5.1.2",
-    "semantic-release": "^15.1.2",
+    "protractor": "~5.3.1",
+    "semantic-release": "^15.1.7",
     "standard-version": "^4.3.0",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "ts-node": "~6.0.1",
+    "tslint": "^5.9.1",
+    "typescript": "~2.8.3",
     "yarn-run-all": "^3.1.1"
   },
   "repository": {
