@@ -3,12 +3,11 @@ import { Movie } from '../movies.model';
 import { MoviesService } from '../movies.service';
 import { MatTableDataSource, MatPaginator, MatSort } from '@angular/material';
 import { SearchMoviesService } from './search-movies.service';
-import { merge } from 'rxjs/Observable/merge';
 import { catchError } from 'rxjs/operators/catchError';
 import { map } from 'rxjs/operators/map';
 import { startWith } from 'rxjs/operators/startWith';
 import { switchMap } from 'rxjs/operators/switchMap';
-import { of as observableOf } from 'rxjs/Observable/of';
+import { of as observableOf } from 'rxjs/observable/of';
 import { HttpClient } from '@angular/common/http';
 import * as UI from '../../shared/store/ui.actions';
 import * as appReducer from '../../app.reducer';
