@@ -1,26 +1,61 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
+import { DomSanitizer } from '@angular/platform-browser';
+import { MatIconRegistry } from '@angular/material';
+import { AuthService } from '../auth/auth.service';
+import { Store } from '@ngrx/store';
 import { HeaderComponent } from './header.component';
 
 describe('HeaderComponent', () => {
-  let component: HeaderComponent;
+  let comp: HeaderComponent;
   let fixture: ComponentFixture<HeaderComponent>;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        declarations: [HeaderComponent]
-      }).compileComponents();
-    })
-  );
-
   beforeEach(() => {
+    const domSanitizerStub = {
+      bypassSecurityTrustResourceUrl: () => ({})
+    };
+    const matIconRegistryStub = {
+      addSvgIcon: () => ({})
+    };
+    const authServiceStub = {
+      logout: () => ({})
+    };
+    const storeStub = {
+      select: () => ({})
+    };
+    TestBed.configureTestingModule({
+      declarations: [HeaderComponent],
+      schemas: [NO_ERRORS_SCHEMA],
+      providers: [
+        { provide: DomSanitizer, useValue: domSanitizerStub },
+        { provide: MatIconRegistry, useValue: matIconRegistryStub },
+        { provide: AuthService, useValue: authServiceStub },
+        { provide: Store, useValue: storeStub }
+      ]
+    });
     fixture = TestBed.createComponent(HeaderComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    comp = fixture.componentInstance;
   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
+  it('can load instance', () => {
+    expect(comp).toBeTruthy();
+  });
+
+  describe('ngOnInit', () => {
+    it('makes expected calls', () => {
+      const storeStub: Store<any> = fixture.debugElement.injector.get(Store);
+      spyOn(storeStub, 'select');
+      comp.ngOnInit();
+      expect(storeStub.select).toHaveBeenCalled();
+    });
+  });
+
+  describe('logout', () => {
+    it('makes expected calls', () => {
+      const authServiceStub: AuthService = fixture.debugElement.injector.get(AuthService);
+      spyOn(authServiceStub, 'logout');
+      comp.logout();
+      expect(authServiceStub.logout).toHaveBeenCalled();
+    });
   });
 });
