@@ -3,7 +3,16 @@ import { Routes, RouterModule } from '@angular/router';
 import { MoviesComponent } from './movies.component';
 import { SearchMoviesComponent } from './search-movies/search-movies.component';
 
-const routes: Routes = [{ path: '', component: MoviesComponent }, { path: 'search', component: SearchMoviesComponent }];
+const routes: Routes = [
+  {
+    path: '',
+    component: MoviesComponent
+  },
+  {
+    path: 'search',
+    component: SearchMoviesComponent
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
